+++ src/app/components/Card/JobCardList.tsx
