+++ src/app/components/Card/View.tsx
