+++ src/app/components/Filter/PopOver.tsx
@@ -2,7 +2,7 @@
 import { createPortal } from "react-dom";
 import { useEffect, useRef, useState } from "react";
 import type { Option } from "./FilterTag";
-import { useIsDesktop } from "../hooks/useIsDesktop";
+import { useIsDesktop } from "../../hooks/useIsDesktop";
 import FilterOptionList from "./FilterOptionList";
 export function PopoverPanel({
   open,
